+++ CHANGELOG.md
@@ -4,8 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## Unreleased
---
+## [1.1.0] - 2021-03-22
+### Added
+- Added new built-in commands
+    - `rm`: Removes files and directories
+    - `mkdir`: Creates directories
+
+### Changed
+- Changed `Third Party Notices.md` markdown
 
 ## [1.0.0] - 2021-03-02
 ### Added